将单位us转换成ns
# message auto-generated for no-merge-commit merge:
merge jingdu into master

将存储在数据库内的时间戳的单位统一为ns

Created-by: insightDeveloper

Match-id-00bf38c620b8ec5c980bc34de4f940b65e3e6ae7

+++ server/src/modules/memory/parser/MemoryDef.h
@@ -13,18 +13,18 @@ namespace Memory {
 
 struct Record {
     std::string component;
-    int64_t timesTamp;
-    int64_t totalAllocated;
-    int64_t totalReserved;
+    double timesTamp;
+    double totalAllocated;
+    double totalReserved;
     std::string deviceType;
 };
 
 struct Operator {
     std::string name;
-    int64_t size;
-    int64_t allocationTime;
-    int64_t releaseTime;
-    int64_t duration;
+    double size;
+    double allocationTime;
+    double releaseTime;
+    double duration;
     std::string deviceType;
 };
